move ContractSeo to it's own file
